Expose the value of "today" in settings
When the initial date is Nothing, clicking on the DatePicker brings up a
pop-up box with the date set to 29/05/1992. However we want this date to
be "today", or any other date we choose.

This exposes the concept of `today` in the settings enabling the default
date to be overridden. If `today` is not set, then `initDate` is still
used, so the current behaviour does not change.